Implements solana transfer.
